detalles indice

+++ finanzas/views.py
@@ -1723,14 +1723,14 @@ def indicelink(request, id_moneda, id_time):
 
             if (dato.proyecto.fecha_i.month - ahora.month) > 0:
                 meses_costo = dato.proyecto.fecha_f.month - dato.proyecto.fecha_i.month
-                meses_costo = meses_costo/2 + dato.proyecto.fecha_i.month - ahora.month
+                meses_costo = int(meses_costo/2 + dato.proyecto.fecha_i.month - ahora.month)
             else:
                 if (dato.proyecto.fecha_f.month - ahora.month) > 0:
-                    meses_costo = (dato.proyecto.fecha_f.month - ahora.month)/2
+                    meses_costo = int((dato.proyecto.fecha_f.month - ahora.month)/2)
                 else:
                     meses_costo = 0           
 
-            meses_ingreso = (dato.proyecto.fecha_f.month - ahora.month)/2
+            meses_ingreso = int((dato.proyecto.fecha_f.month - ahora.month)/2)
 
             if meses_costo:
                 array_costo = np.zeros(meses_costo, dtype = int)
